BIGC-630: removed order-placed.js and refactored both themes to do what was done in order-placed before. Changed Readme.md
