+++ webpack.config.js
@@ -3,7 +3,6 @@ const webpack = require('webpack');
 
 const webpackConfiguration = {
   entry: {
-    'src/order-placed': './src/order-placed.js',
     'src/themes/cornerstone': './src/themes/cornerstone.js',
     'src/themes/fortune': './src/themes/fortune.js',
   },
